Move quick reply definitions
